moves widgets to eto folder
