+++ app/components/eto/dashboard/bookBuildingWidget/BookBuildingWidget.tsx
@@ -4,10 +4,10 @@ import { FormattedMessage } from "react-intl-phraseapp";
 import { Col } from "reactstrap";
 import { compose } from "redux";
 
-import { appConnect } from "../../../store";
-import { IIntlProps, injectIntlHelpers } from "../../../utils/injectIntlHelpers";
-import { ButtonArrowRight } from "../../shared/Buttons";
-import { Panel } from "../../shared/Panel";
+import { appConnect } from "../../../../store";
+import { IIntlProps, injectIntlHelpers } from "../../../../utils/injectIntlHelpers";
+import { ButtonArrowRight } from "../../../shared/Buttons";
+import { Panel } from "../../../shared/Panel";
 
 import * as styles from "./BookBuildingWidget.module.scss";
