+++ app/components/eto/dashboard/choosePreEtoDateWidget/ChoosePreEtoDateWidget.stories.tsx
@@ -1,7 +1,8 @@
 import { storiesOf } from "@storybook/react";
 import * as React from "react";
 
-import { dummyIntl } from "../../../utils/injectIntlHelpers.fixtures";
+import { dummyIntl } from "../../../../utils/injectIntlHelpers.fixtures";
+
 import { ChoosePreEtoDateWidgetComponent } from "./ChoosePreEtoDateWidget";
 
 storiesOf("ChoosePreEtoDateWidget", module).add("default", () => (
