+++ app/components/eto/dashboard/submitProposalWidget/SubmitProposalWidget.tsx
@@ -4,11 +4,11 @@ import { FormattedMessage } from "react-intl-phraseapp";
 import { Col } from "reactstrap";
 import { compose } from "redux";
 
-import { actions } from "../../../modules/actions";
-import { appConnect } from "../../../store";
-import { IIntlProps, injectIntlHelpers } from "../../../utils/injectIntlHelpers";
-import { ButtonArrowRight } from "../../shared/Buttons";
-import { Panel } from "../../shared/Panel";
+import { actions } from '../../../../modules/actions';
+import { appConnect } from "../../../../store";
+import { IIntlProps, injectIntlHelpers } from "../../../../utils/injectIntlHelpers";
+import { ButtonArrowRight } from "../../../shared/Buttons";
+import { Panel } from "../../../shared/Panel";
 
 import * as styles from "./SubmitProposalWidget.module.scss";
