+++ app/components/eto/dashboard/uploadProspectusWidget/UploadProspectusWidget.tsx
@@ -3,13 +3,14 @@ import * as React from "react";
 import { FormattedMessage } from "react-intl-phraseapp";
 import { Link } from "react-router-dom";
 import { Col } from "reactstrap";
-import { appRoutes } from "../../appRoutes";
 
-import { injectIntlHelpers } from "../../../utils/injectIntlHelpers";
-import { ButtonArrowRight } from "../../shared/Buttons";
-import { Panel } from "../../shared/Panel";
+import { injectIntlHelpers } from "../../../../utils/injectIntlHelpers";
+import { appRoutes } from "../../../appRoutes";
+import { ButtonArrowRight } from "../../../shared/Buttons";
+import { Panel } from "../../../shared/Panel";
 
 import * as styles from "./UploadProspectusWidget.module.scss";
+
 
 export const UploadProspectusWidget = injectIntlHelpers(({ intl: { formatIntlMessage } }) => {
   return (
